+++ Frontend/src/lib/images/index.ts
@@ -2,8 +2,8 @@ export {default as icon} from './logo/icon.svg';
 export {default as logo} from './logo/logo-white-text.svg';
 export {default as logoBesideText} from './logo/logo-beside-text.svg';
 
-export {default as github} from './links/github.svg';
 export {default as simplex} from './links/simplex.svg';
+export {default as github} from './links/github.svg';
 export {default as tor} from './links/tor.svg';
 
 export {default as info} from './status/info.svg';
@@ -16,7 +16,9 @@ export {default as itCrowdGif} from './404/IT-Crowd.gif';
 export {default as satisfaction} from './homepage/satisfaction.png';
 export {default as background} from './homepage/background.png';
 export {default as dashboard} from './homepage/dashboard.png';
-export {default as hideout} from './homepage/hideout.svg';
+
+export {default as registration} from './homepage/benefits/registration.svg';
+export {default as management} from './homepage/benefits/management.svg';
 
 export {default as cog} from './homepage/features/cog.svg';
 export {default as tag} from './homepage/features/tag.svg';
@@ -27,13 +29,14 @@ export {default as customer} from './homepage/features/customer.svg';
 export {default as terminal} from './homepage/features/terminal.svg';
 export {default as openSource} from './homepage/features/open-source.svg';
 
-export {default as card} from './homepage/cards/card.png';
-export {default as cloud} from './homepage/cards/cloud.png';
-export {default as email} from './homepage/cards/email.png';
-export {default as globe} from './homepage/cards/globe.png';
-export {default as money} from './homepage/cards/money.png';
-export {default as tower} from './homepage/cards/tower.png';
-export {default as router} from './homepage/cards/router.png';
-export {default as server} from './homepage/cards/server.png';
-export {default as message} from './homepage/cards/message.png';
-export {default as identity} from './homepage/cards/identity.png';
+export {default as card} from './homepage/capabilities/card.png';
+export {default as cloud} from './homepage/capabilities/cloud.png';
+export {default as email} from './homepage/capabilities/email.png';
+export {default as globe} from './homepage/capabilities/globe.png';
+export {default as money} from './homepage/capabilities/money.png';
+export {default as tower} from './homepage/capabilities/tower.png';
+export {default as router} from './homepage/capabilities/router.png';
+export {default as server} from './homepage/capabilities/server.png';
+export {default as hideout} from './homepage/capabilities/hideout.svg';
+export {default as message} from './homepage/capabilities/message.png';
+export {default as identity} from './homepage/capabilities/identity.png';